Cleanup Templates Code
