+++ src/components/templates/index.tsx
@@ -1,1 +1,3 @@
-export * from './Regular';
+export {default as RegularTemplate} from './RegularTemplate';
+export {default as RegularRowOne} from './RegularRowOne';
+export {default as RegularRowTwo} from './RegularRowTwo';
